tests/lnprototest/pyproject.toml
[tool.poetry]
name = "lampo-lnprototest"
version = "0.0.1"
description = "lnprototest runner for lampo"
authors = ["Vincenzo Palazzo <vincenzopalazzodev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
lnprototest = "^0.0.3"
lampo-py = "0.0.1a3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pyln-bolt4 = "^1.0.222"
pyln-bolt2 = "^1.0.222"
pyln-bolt1 = "^1.0.222"
pyln-client = "^0.12.0"
crc32c = "^2.2.post0"
# We accidentally published version 1.0.186 instead of 1.0.2.186. That
# version is now yanked by caches remain, so this is a temporary fix.
pyln-bolt7 = "^1.0.246"
pyln-proto = "^0.12.0"
python-bitcoinlib = "^0.11.2"
black = "^22.1.0"
flake8 = "^4.0.1"
pytest-xdist = "^3.1.0"
mypy = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
